Corrected an error where caused by last update where the scripts would not try to gain root access
Added logic into remoteHack.js to begin by populating the list of hosts to use
in case it is empty. Then, before starting the hacks on the hosts, it runs
populateHackHost.js on the known hack hosts. This in effect makes it
to where all hack hosts are constantly grabbing nearby hosts. This did require
adding a delay to the script of 2 seconds.