Renamed test.js to infiltrationTest.js to be more descriptive
Finished the sortByMoney function in sortHackTargets.js
Used a more efficient method to perform both sorts

Altered the determineHackTarget.js script to use the sort method from sortHackTargets.js
on a filtered list of only targets that are less than or equal to the player hack
level, where the player has root access, and which have a max money of more than zero.
Rahter than return the server with the closest hack level to the player, it now returns the
most lucrative server that the player can hack.